Website Updates
The file names were updated to index.html, dr_info.html, dr_faq.html. The meta data was set on both the info and index page. Both the index and info page share the same header information. Diane’s Run was put into a header 1. The images for the top of the page was added for both index and info. The main text was added to both the index and info page. The footer was added to the index and info page. Section elements were implemented on both pages in order to section off the page.