Website Updated Finished
The base.css and dr_layout3.css was linked to the dr_faq page. The header and footer of index.html was copied to the dr_faq.html page. An unordered list was added to the dr_faq.html page to act as a list for the questions. A link was added to take viewers exactly to that questions. Each of the questions were put into the h2 elements with an id attribute. The id attributes contained faq(number) so that it could be linked up. A got to page top line was added with an id attribute of #pagetop. An image was added to the top of the faq page. In the index.html a link was added for the email in the second paragraph. That link auto filled the subject line. Also in the index.html page accommodations was linked up with dr_faq.html#faq13.